requirements.txt
# Install FastAPI and dependencies
fastapi
uvicorn
python-multipart


# Install LangGraph and related libraries
langgraph
langchain
langchain-community
langsmith
python-dotenv
langchain-google-genai
langchain-tavily

# Install libraries for vector databases and embeddings
chromadb
sentence_transformers
tqdm
PyMuPDF
slowapi

# Install libraries for Redis and other utilities
redis
